api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
slowapi==0.1.9
httpx==0.27.0
elasticsearch==8.13.2 
orjson==3.10.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
fhir-resources==7.1.0
tqdm




